+++ src/addons/vue-template.ts
@@ -48,9 +48,11 @@ export const vueTemplateAddon = (): Addon => ({
     const imports = await this.getImports()
     const items = imports
       .map((i) => {
+        if (i.type) { return '' }
         const from = options?.resolvePath?.(i) || i.from
         return `readonly ${i.as}: UnwrapRef<typeof import('${from}')${i.name !== '*' ? `['${i.name}']` : ''}>`
       })
+      .filter(Boolean)
       .sort()
     return dts +
 `