+++ src/presets/vue-router-composables.ts
@@ -1,0 +1,12 @@
+import { defineUnimportPreset } from '../utils'
+
+export default defineUnimportPreset({
+  from: 'vue-router/composables',
+  imports: [
+    'useRouter',
+    'useRoute',
+    'useLink',
+    'onBeforeRouteLeave',
+    'onBeforeRouteUpdate'
+  ]
+})