+++ src/types.ts
@@ -26,6 +26,10 @@ export interface ImportCommon {
     /** Additional metadata */
     [key: string]: any
   }
+  /**
+   * If this import is a type import
+   */
+  type?: boolean
 }
 
 export interface Import extends ImportCommon {
@@ -35,7 +39,7 @@ export interface Import extends ImportCommon {
   as?: ImportName
 }
 
-export type PresetImport = ImportName | [name: ImportName, as?: ImportName, from?: ModuleId] | Exclude<Import, 'from'>
+export type PresetImport = Omit<Import, 'from'> | ImportName | [name: ImportName, as?: ImportName, from?: ModuleId]
 
 export interface InlinePreset extends ImportCommon {
   imports: (PresetImport | InlinePreset)[]
@@ -80,6 +84,8 @@ export interface UnimportContext {
   getImportMap(): Promise<Map<string, Import>>
   getMetadata(): UnimportMeta | undefined
 
+  replaceImports(imports: UnimportOptions['imports']): Promise<Import[]>
+
   invalidate(): void
   resolveId(id: string, parentId?: string): Thenable<string | null | undefined | void>
 }
@@ -103,7 +109,7 @@ export interface AddonsOptions {
   vueTemplate?: boolean
 }
 
-export interface UnimportOptions {
+export interface UnimportOptions extends Pick<InjectImportsOptions, 'injectAtEnd' | 'mergeExisting'> {
   /**
    * Auto import items
    */
@@ -213,6 +219,12 @@ export interface TypeDeclarationOptions {
    * @default true
    */
   exportHelper?: boolean
+  /**
+   * Auto-import for type exports
+   *
+   * @default true
+   */
+  typeReExports?: boolean
 }
 
 export interface InjectImportsOptions {
@@ -240,6 +252,13 @@ export interface InjectImportsOptions {
 
   /** @deprecated use `virtualImports` instead */
   transformVirtualImoports?: boolean
+
+  /**
+   * Inject the imports at the end of other imports
+   *
+   * @default false
+   */
+  injectAtEnd?: boolean
 }
 
 export type Thenable<T> = Promise<T> | T