+++ test/dts.test.ts
@@ -12,7 +12,20 @@ test('dts', async () => {
     presets: [
       {
         from: 'vue',
-        imports: ['ref', 'reactive', 'computed', 'toRefs']
+        imports: [
+          'ref',
+          'reactive',
+          'computed',
+          'toRefs',
+          {
+            name: 'Ref',
+            type: true
+          },
+          {
+            name: 'ComputedRef',
+            type: true
+          }
+        ]
       },
       {
         from: 'three',
@@ -24,7 +37,13 @@ test('dts', async () => {
       },
       {
         from: 'jquery',
-        imports: ['$']
+        imports: [
+          '$',
+          {
+            name: 'JQuery',
+            type: true
+          }
+        ]
       }
     ],
     dirs: [
@@ -37,7 +56,9 @@ test('dts', async () => {
 
   await init()
 
-  expect((await generateTypeDeclarations()).replaceAll(cwd, '<root>'))
+  expect(
+    (await generateTypeDeclarations()).replaceAll(cwd, '<root>')
+  )
     .toMatchInlineSnapshot(`
       "export {}
       declare global {
@@ -63,6 +84,13 @@ test('dts', async () => {
         const useEffect: typeof import('react')['useEffect']
         const useRef: typeof import('react')['useRef']
         const useState: typeof import('react')['useState']
+      }
+      // for type re-export
+      declare global {
+        // @ts-ignore
+        export type { Ref, ComputedRef } from 'vue'
+        // @ts-ignore
+        export type { JQuery } from 'jquery'
       }"
     `)
 })